services/ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings>=2.0
requests
google.generativeai